feat: write operation API 어드민 체크 및 접근URL 변경 - checkAdminAccess로 어드민 계정 검증 - 메서드 호출이 허용되는지 여부를 결정하기 위해 평가할 메서드 액세스 제어 식 @PreAuthorize ADMIN으로 추가 - spring security를 통한 인가에 대한 접근도 ADMIN으로 가능하도록 URL /api/admin/..으로 수정

+++ src/main/java/com/youngcamp/server/controller/AnnouncementController.java
@@ -7,10 +7,13 @@ import com.youngcamp.server.dto.AnnouncementResponse.AnnouncementEditResponse;
 import com.youngcamp.server.dto.AnnouncementResponse.AnnouncementGetDetailResponse;
 import com.youngcamp.server.dto.AnnouncementResponse.AnnouncementGetResponse;
 import com.youngcamp.server.dto.AnnouncementResponse.AnnouncementPostResponse;
+import com.youngcamp.server.service.AdminChecker;
 import com.youngcamp.server.service.AnnouncementService;
 import com.youngcamp.server.utils.SuccessResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -20,13 +23,16 @@ import java.util.List;
 public class AnnouncementController {
 
     private final AnnouncementService announcementService;
+    private final AdminChecker adminChecker;
 
     @Operation(
             summary = "공지사항 등록 API",
             description = "공지사항을 등록합니다."
     )
-    @PostMapping("/api/v1/announcements")
+    @PreAuthorize("hasRole('ADMIN')")
+    @PostMapping("/api/admin/v1/announcements")
     public SuccessResponse<AnnouncementPostResponse> postAnnouncement(@RequestBody AnnouncementPostRequest request) {
+        checkAdminAccess();
         AnnouncementPostResponse result = announcementService.addAnnouncement(request);
         return new SuccessResponse<>("Request processed successfully", result);
     }
@@ -35,8 +41,10 @@ public class AnnouncementController {
             summary = "공지사항 삭제 API",
             description = "공지사항 ID값을 List로 넘겨 받아 Batch삭제 합니다."
     )
-    @DeleteMapping("/api/v1/announcements")
+    @PreAuthorize("hasRole('ADMIN')")
+    @DeleteMapping("/api/admin/v1/announcements")
     public SuccessResponse<?> deleteAnnouncements(@RequestBody AnnouncementDeleteRequest request) {
+        checkAdminAccess();
         announcementService.deleteAnnouncement(request);
         return new SuccessResponse<>("Request processed successfully", null); //TODO
     }
@@ -65,11 +73,13 @@ public class AnnouncementController {
             summary = "공지사항 수정 API",
             description = "공지사항 ID값으로 특정 공지 사항 내용을 수정합니다."
     )
-    @PatchMapping("/api/v1/announcements/{announcementId}")
+    @PreAuthorize("hasRole('ADMIN')")
+    @PatchMapping("/api/admin/v1/announcements/{announcementId}")
     public SuccessResponse<AnnouncementEditResponse> editAnnouncement(
             @PathVariable(name = "announcementId") Long announcementId,
             @RequestBody AnnouncementEditRequest request
     ) {
+        checkAdminAccess();
         AnnouncementEditResponse result = announcementService.editAnnouncement(announcementId, request);
         return new SuccessResponse<>("Request processed successfully", result);
     }
@@ -82,5 +92,11 @@ public class AnnouncementController {
     public SuccessResponse<List<AnnouncementGetResponse>> searchAnnouncements(@RequestParam(name = "keyword") String keyword) {
         List<AnnouncementGetResponse> result = announcementService.searchAnnouncements(keyword);
         return new SuccessResponse<>("Request processed successfully", result);
+    }
+
+    private void checkAdminAccess() {
+        if (!adminChecker.isAdmin()) {
+            throw new AccessDeniedException("Access is denied. Admin role required");
+        }
     }
 }
